+++ app/controllers/v2/base_controller.rb
@@ -1,0 +1,12 @@
+module V2
+  class BaseController < ApplicationController
+    def edition_params
+      params.require("edition").permit(
+        :description,
+        :lead_organisation_id,
+        :instructions_to_publishers,
+        :title,
+      ).merge(creator: current_user)
+    end
+  end
+end
